Handle pre-existing root ini files

+++ src/ComposerInstaller.php
@@ -13,9 +13,7 @@ class ComposerInstaller
 
         // Determine if ORM ini already exists.
         $iniPath = realpath($vendorDir.$ds.'..'.$ds.'config').$ds;
-        if (file_exists($iniPath.'GCWorld_ORM.ini')) {
-            return true;    // Already Exists.
-        }
+
         if (!is_dir($iniPath)) {
             @mkdir($iniPath);
             if (!is_dir($iniPath)) {
@@ -23,8 +21,10 @@ class ComposerInstaller
                 return false;   // Silently Fail.
             }
         }
-        $example = file_get_contents($myDir.$ds.'..'.$ds.'config'.$ds.'config.example.ini');
-        file_put_contents($iniPath.'GCWorld_ORM.ini', $example);
+        if (!file_exists($iniPath.'GCWorld_ORM.ini')) {
+            $example = file_get_contents($myDir.$ds.'..'.$ds.'config'.$ds.'config.example.ini');
+            file_put_contents($iniPath.'GCWorld_ORM.ini', $example);
+        }
         file_put_contents($myDir.$ds.'..'.$ds.'config'.$ds.'config.ini', 'config_path='.$iniPath.'GCWorld_ORM.ini');
         return true;
     }
